+++ packages/uploadthing/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @yukinu/uploadthing
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`78ac45e`](https://github.com/tiesen243/yukinu/commit/78ac45e18685ff6c3d8f6b4378b559881709e937)]:
+  - @yukinu/validators@0.9.1
+  - @yukinu/auth@0.9.1
+
 ## 0.3.0
 
 ### Minor Changes